pull pause mgr out of api
