+++ api/src/opentrons/hardware_control/api.py
@@ -11,7 +11,7 @@ from opentrons_shared_data.pipette import name_config
 from opentrons import types as top_types
 from opentrons.util import linal
 from functools import lru_cache
-from opentrons.config import robot_configs, feature_flags as ff
+from opentrons.config import robot_configs
 from opentrons.config.types import RobotConfig
 
 from .util import (
@@ -24,11 +24,12 @@ from .constants import (SHAKE_OFF_TIPS_SPEED, SHAKE_OFF_TIPS_DROP_DISTANCE,
                         SHAKE_OFF_TIPS_PICKUP_DISTANCE,
                         DROP_TIP_RELEASE_DISTANCE)
 from .execution_manager import ExecutionManager
+from .pause_manager import PauseManager
 from .types import (Axis, HardwareAPILike, CriticalPoint,
                     MustHomeError, NoTipAttachedError, DoorState,
                     DoorStateNotification, PipettePair, TipAttachedError,
                     HardwareAction, PairedPipetteConfigValueError,
-                    MotionChecks, PauseType, PauseResumeError)
+                    MotionChecks, PauseType)
 from . import modules, robot_calibration as rb_cal
 
 if TYPE_CHECKING:
@@ -43,44 +44,6 @@ mod_log = logging.getLogger(__name__)
 
 InstrumentsByMount = Dict[top_types.Mount, Optional[Pipette]]
 PipetteHandlingData = Tuple[Pipette, top_types.Mount]
-
-
-class PauseManager:
-    """ This class determines whether or not the hardware controller should
-    pause or resume by evaluating the pause and resume types. The use of two
-    pause types are used to separate the delay resume (triggered when the delay
-    timer runs out) and the pause resume (trigged by user via the app).
-    """
-
-    def __init__(self, door_state: DoorState) -> None:
-        self.queue: List[PauseType] = []
-        self._blocked_by_door = self._evaluate_door_state(door_state)
-
-    @property
-    def blocked_by_door(self) -> bool:
-        return self._blocked_by_door
-
-    def _evaluate_door_state(self, door_state: DoorState) -> bool:
-        if ff.enable_door_safety_switch():
-            return door_state is DoorState.OPEN
-        return False
-
-    def set_door(self, door_state: DoorState):
-        self._blocked_by_door = self._evaluate_door_state(door_state)
-
-    def resume(self, pause_type: PauseType):
-        # door should be closed before a resume from the app can be received
-        if self._blocked_by_door and pause_type is PauseType.PAUSE:
-            raise PauseResumeError
-        if pause_type in self.queue:
-            self.queue.remove(pause_type)
-
-    def pause(self, pause_type: PauseType):
-        if pause_type not in self.queue:
-            self.queue.append(pause_type)
-
-    def reset(self):
-        self.queue = []
 
 
 class API(HardwareAPILike):
